Move DebuggingWorkspaceService to EditorFeatures
