+++ src/EditorFeatures/Core/Implementation/Debugging/IDebuggingWorkspaceService.cs
@@ -2,8 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-#nullable disable
-
 using System;
 using Microsoft.CodeAnalysis.Host;
 
@@ -11,7 +9,7 @@ namespace Microsoft.CodeAnalysis.Debugging
 {
     internal interface IDebuggingWorkspaceService : IWorkspaceService
     {
-        event EventHandler<DebuggingStateChangedEventArgs> BeforeDebuggingStateChanged;
+        event EventHandler<DebuggingStateChangedEventArgs>? BeforeDebuggingStateChanged;
 
         DebuggingState CurrentDebuggingState { get; }
